requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# PDF Processing
PyMuPDF==1.23.21
pdfplumber==0.11.0

# Citation Extraction
eyecite==2.6.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Email
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
emails==0.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
